Changed back to layoutbased structure
