+++ src/MudBlazor.Docs/Pages/Mud/_Imports.razor
@@ -1,0 +1,1 @@
+﻿@layout DocsMoreLayout
