+++ src/MudBlazor.Docs/Shared/DocsLayout.razor.cs
@@ -9,37 +9,11 @@ namespace MudBlazor.Docs.Shared;
 
 public partial class DocsLayout : LayoutComponentBase
 {
-    [CascadingParameter] private MainLayout MainData { get; set; }
-    
     private NavMenu _navMenuRef;
-    private bool _drawerOpen = true;
-    private bool _topMenuOpen = false;
-    
-    protected override void OnInitialized()
-    {
-         MainData.SetBaseTheme(Theme.DocsTheme());
-    }
-    
+
     protected override void OnAfterRender(bool firstRender)
     {
         //refresh nav menu because no parameters change in nav menu but internal data does
         _navMenuRef.Refresh();
-    }
-
-    private void ToggleDrawer()
-    {
-        _drawerOpen = !_drawerOpen;
-    }
-
-    private void OpenTopMenu()
-    {
-        _topMenuOpen = true;
-    }
-
-    private void OnDrawerOpenChanged(bool value)
-    {
-        _topMenuOpen = false;
-        _drawerOpen = value;
-        StateHasChanged();
     }
 }
