+++ src/MudBlazor.Docs/Shared/NavMenu.razor
@@ -1,76 +1,43 @@
 @using System.Linq
 @using MudBlazor.Docs.Extensions
 
-<MudNavMenu Color="Color.Primary" Rounded="true" Dense="true" Margin="Margin.Dense" Class="pa-2 overflow-auto mb-3">
-    @if (MainData.GetDocsBasePage() == DocsBasePage.GettingStarted)
+<MudText Typo="Typo.subtitle1" Class="mt-1 mb-2 ml-2 d-flex d-md-none"><b>Docs</b></MudText>
+<MudNavGroup Title="Components" Icon="@Icons.Material.Filled.Dashboard" Expanded="@(_section == "components")" ExpandIcon="@Icons.Material.Filled.ExpandMore">
+    @foreach (var item in MenuService.Components)
     {
-        <MudText Typo="Typo.subtitle1" Class="mt-1 mb-2 ml-2 d-flex d-md-none"><b>Getting Started</b></MudText>
-        <MudNavLink Href="/getting-started/installation">Installation</MudNavLink>
-        <MudNavLink Href="/getting-started/layouts">Layouts</MudNavLink>
-        <MudNavLink Href="/getting-started/usage">Usage</MudNavLink>
-        <MudNavLink Href="/getting-started/wireframes">Wireframes</MudNavLink>
-    }
-
-    @if (MainData.GetDocsBasePage() == DocsBasePage.DiscoverMore)
-    {
-        <MudText Typo="Typo.subtitle1" Class="mt-1 mb-2 ml-2 d-flex d-md-none"><b>Discover More</b></MudText>
-        <MudNavLink Href="/mud/introduction">What is MudBlazor?</MudNavLink>
-        <MudNavGroup Title="Community" Expanded="true" HideExpandIcon="true">
-            <MudNavLink Href="/mud/community/getting-help">Getting Help</MudNavLink>
-            <MudNavLink Href="/mud/community/reporting-bugs">Reporting Bugs</MudNavLink>
-            <MudNavLink Href="/mud/community/contribution">Contribution</MudNavLink>
-        </MudNavGroup>
-        <MudNavGroup Title="Project" Expanded="true" HideExpandIcon="true">
-            <MudNavLink Href="/mud/announcements">Announcements</MudNavLink>
-            <MudNavLink Href="/mud/project/how-it-started">How it started</MudNavLink>
-            <MudNavLink Href="/mud/project/releases">Releases</MudNavLink>
-            <MudNavLink Href="/mud/project/roadmap">Roadmap</MudNavLink>
-            <MudNavLink Href="/mud/project/sponsor">Sponsors & Backers</MudNavLink>
-            <MudNavLink Href="/mud/project/team">Team & Contributors</MudNavLink>
-        </MudNavGroup>
-    }
-
-    @if (MainData.GetDocsBasePage() == DocsBasePage.Docs)
-    {
-        <MudText Typo="Typo.subtitle1" Class="mt-1 mb-2 ml-2 d-flex d-md-none"><b>Docs</b></MudText>
-        <MudNavGroup Title="Components" Icon="@Icons.Material.Filled.Dashboard" Expanded="@(_section == "components")" ExpandIcon="@Icons.Material.Filled.ExpandMore">
-            @foreach (var item in MenuService.Components)
-            {
-                if (item.IsNavGroup)
+        if (item.IsNavGroup)
+        {
+            <MudNavGroup Title="@item.Name" Expanded="@(IsSubGroupExpanded(item))" ExpandIcon="@Icons.Material.Filled.ExpandMore">
+                @foreach (var subItem in item.GroupComponents)
                 {
-                    <MudNavGroup Title="@item.Name" Expanded="@(IsSubGroupExpanded(item))" ExpandIcon="@Icons.Material.Filled.ExpandMore">
-                        @foreach (var subItem in item.GroupComponents)
-                        {
-                            string href = $"components/{subItem.Link}";
-                            <MudNavLink Href="@href">@subItem.Name</MudNavLink>
-                        }
-                    </MudNavGroup>
+                    string href = $"components/{subItem.Link}";
+                    <MudNavLink Href="@href">@subItem.Name</MudNavLink>
                 }
-                else
-                {
-                    string href = $"components/{item.Link}";
-                    <MudNavLink Href="@href">@item.Name</MudNavLink>
-                }
-            }
-        </MudNavGroup>
-        <MudNavGroup Title="API" Icon="@Icons.Material.Filled.Api" Expanded="@(_section == "api")" ExpandIcon="@Icons.Material.Filled.ExpandMore">
-            @foreach (var item in MenuService.Api.OrderBy(x => x.Name))
-            {
-                <MudNavLink Href="@ApiLink.GetApiLinkFor(item.Type)">@item.Name</MudNavLink>
-            }
-        </MudNavGroup>
-        <MudNavGroup Title="Features" Icon="@Icons.Material.Filled.DeveloperBoard" Expanded="@(_section == "features")" ExpandIcon="@Icons.Material.Filled.ExpandMore">
-            @foreach (var link in MenuService.Features)
-            {
-                <MudNavLink Href="@link.Href">@link.Title</MudNavLink>
-            }
-        </MudNavGroup>
-        <MudNavGroup Title="Customization" Icon="@Icons.Material.Filled.Handyman" Expanded="@(_section == "customization")" ExpandIcon="@Icons.Material.Filled.ExpandMore">
-            <MudNavLink Href="customization/default-theme">Default Theme</MudNavLink>
-            @foreach (var link in MenuService.Customization)
-            {
-                <MudNavLink Href="@link.Href">@link.Title</MudNavLink>
-            }
-        </MudNavGroup>
+            </MudNavGroup>
+        }
+        else
+        {
+            string href = $"components/{item.Link}";
+            <MudNavLink Href="@href">@item.Name</MudNavLink>
+        }
     }
-</MudNavMenu>
+</MudNavGroup>
+<MudNavGroup Title="API" Icon="@Icons.Material.Filled.Api" Expanded="@(_section == "api")" ExpandIcon="@Icons.Material.Filled.ExpandMore">
+    @foreach (var item in MenuService.Api.OrderBy(x => x.Name))
+    {
+        <MudNavLink Href="@ApiLink.GetApiLinkFor(item.Type)">@item.Name</MudNavLink>
+    }
+</MudNavGroup>
+<MudNavGroup Title="Features" Icon="@Icons.Material.Filled.DeveloperBoard" Expanded="@(_section == "features")" ExpandIcon="@Icons.Material.Filled.ExpandMore">
+    @foreach (var link in MenuService.Features)
+    {
+        <MudNavLink Href="@link.Href">@link.Title</MudNavLink>
+    }
+</MudNavGroup>
+<MudNavGroup Title="Customization" Icon="@Icons.Material.Filled.Handyman" Expanded="@(_section == "customization")" ExpandIcon="@Icons.Material.Filled.ExpandMore">
+    <MudNavLink Href="customization/default-theme">Default Theme</MudNavLink>
+    @foreach (var link in MenuService.Customization)
+    {
+        <MudNavLink Href="@link.Href">@link.Title</MudNavLink>
+    }
+</MudNavGroup>
